decouple LineLengthTransformer
