+++ packages/CodingStandard/src/Fixer/LineLength/BreakNewInstanceArgumentsFixer.php
@@ -9,51 +9,19 @@ use PhpCsFixer\FixerDefinition\FixerDefinition;
 use PhpCsFixer\FixerDefinition\FixerDefinitionInterface;
 use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
-use PhpCsFixer\WhitespacesFixerConfig;
 use SplFileInfo;
-use Symplify\TokenRunner\Analyzer\FixerAnalyzer\IndentDetector;
-use Symplify\TokenRunner\Configuration\Configuration;
+use Symplify\TokenRunner\Transformer\FixerTransformer\LineLengthTransformer;
 
 final class BreakNewInstanceArgumentsFixer implements DefinedFixerInterface
 {
     /**
-     * @var WhitespacesFixerConfig
+     * @var LineLengthTransformer
      */
-    private $whitespacesFixerConfig;
+    private $lineLengthTransformer;
 
-    /**
-     * @var IndentDetector
-     */
-    private $indentDetector;
-
-    /**
-     * @var string
-     */
-    private $indentWhitespace;
-
-    /**
-     * @var string
-     */
-    private $newlineIndentWhitespace;
-
-    /**
-     * @var string
-     */
-    private $closingBracketNewlineIndentWhitespace;
-
-    /**
-     * @var Configuration
-     */
-    private $configuration;
-
-    public function __construct(
-        Configuration $configuration,
-        IndentDetector $indentDetector,
-        WhitespacesFixerConfig $whitespacesFixerConfig
-    ) {
-        $this->indentDetector = $indentDetector;
-        $this->whitespacesFixerConfig = $whitespacesFixerConfig;
-        $this->configuration = $configuration;
+    public function __construct(LineLengthTransformer $lineLengthTransformer)
+    {
+        $this->lineLengthTransformer = $lineLengthTransformer;
     }
 
     public function getDefinition(): FixerDefinitionInterface
@@ -91,19 +59,16 @@ $someObject = new SomeClass($superLongArguments, $anotherLongArguments, $andLitt
             // the type Tokens::BLOCK_TYPE_PARENTHESIS_BRACE manually, I never get it from the name
             $endBracketPosition = $tokens->findBlockEnd(Tokens::BLOCK_TYPE_PARENTHESIS_BRACE, $startBracketPosition);
 
-            // no arguments => skip
-            if (($endBracketPosition - $startBracketPosition) <= 1) {
+            if ($this->shouldSkip($tokens, $startBracketPosition, $endBracketPosition)) {
                 continue;
             }
 
-            // nowdoc => skip
-            $nextTokenPosition = $tokens->getNextMeaningfulToken($startBracketPosition);
-            $nextToken = $tokens[$nextTokenPosition];
-            if (Strings::startsWith($nextToken->getContent(), '<<<')) {
-                continue;
-            }
-
-            $this->fixStartPositionToEndPosition($startBracketPosition, $endBracketPosition, $tokens, $position);
+            $this->lineLengthTransformer->fixStartPositionToEndPosition(
+                $startBracketPosition,
+                $endBracketPosition,
+                $tokens,
+                $position
+            );
         }
     }
 
@@ -130,160 +95,17 @@ $someObject = new SomeClass($superLongArguments, $anotherLongArguments, $andLitt
         return true;
     }
 
-    /**
-     * First steps to general fixer
-     */
-    private function fixStartPositionToEndPosition(
-        int $startPosition,
-        int $endPosition,
-        Tokens $tokens,
-        int $currentPosition
-    ): void {
-        // @todo automate in some way
-        $this->prepareIndentWhitespaces($tokens, $startPosition);
-
-        $firstLineLength = $this->getFirstLineLength($startPosition, $tokens);
-        if ($firstLineLength > $this->configuration->getMaxLineLength()) {
-            $this->breakItems($startPosition, $endPosition, $tokens);
-            return;
-        }
-
-        $fullLineLength = $this->getLengthFromStartEnd($startPosition, $endPosition, $tokens);
-        if ($fullLineLength <= $this->configuration->getMaxLineLength()) {
-            $this->inlineItems($startPosition, $endPosition, $tokens, $currentPosition);
-            return;
-        }
-    }
-
-    private function prepareIndentWhitespaces(Tokens $tokens, int $arrayStartIndex): void
+    private function shouldSkip(Tokens $tokens, int $startBracketPosition, int $endBracketPosition): bool
     {
-        $indentLevel = $this->indentDetector->detectOnPosition(
-            $tokens,
-            $arrayStartIndex,
-            $this->whitespacesFixerConfig
-        );
-        $indentWhitespace = $this->whitespacesFixerConfig->getIndent();
-        $lineEnding = $this->whitespacesFixerConfig->getLineEnding();
-
-        $this->indentWhitespace = str_repeat($indentWhitespace, $indentLevel + 1);
-        $this->closingBracketNewlineIndentWhitespace = $lineEnding . str_repeat($indentWhitespace, $indentLevel);
-        $this->newlineIndentWhitespace = $lineEnding . $this->indentWhitespace;
-    }
-
-    /**
-     * @abstractable
-     */
-    private function getFirstLineLength(int $startPosition, Tokens $tokens): int
-    {
-        $lineLength = 0;
-
-        // compute from here to start of line
-        $currentPosition = $startPosition;
-        while (! Strings::startsWith($tokens[$currentPosition]->getContent(), PHP_EOL)) {
-            $lineLength += strlen($tokens[$currentPosition]->getContent());
-            --$currentPosition;
+        // no arguments => skip
+        if (($endBracketPosition - $startBracketPosition) <= 1) {
+            return true;
         }
 
-        $currentToken = $tokens[$currentPosition];
+        // nowdoc => skip
+        $nextTokenPosition = $tokens->getNextMeaningfulToken($startBracketPosition);
+        $nextToken = $tokens[$nextTokenPosition];
 
-        // includes indent in the beginning
-        $lineLength += strlen($currentToken->getContent());
-
-        // minus end of lines, do not count PHP_EOL as characters
-        $endOfLineCount = substr_count($currentToken->getContent(), PHP_EOL);
-        $lineLength -= $endOfLineCount;
-
-        // compute from here to end of line
-        $currentPosition = $startPosition + 1;
-        while (! Strings::startsWith($tokens[$currentPosition]->getContent(), PHP_EOL)) {
-            $lineLength += strlen($tokens[$currentPosition]->getContent());
-            ++$currentPosition;
-        }
-
-        return $lineLength;
-    }
-
-    /**
-     * @abstractable
-     */
-    private function breakItems(int $startPosition, int $endPosition, Tokens $tokens): void
-    {
-        // 1. break after arguments opening
-        $tokens->ensureWhitespaceAtIndex($startPosition + 1, 0, $this->newlineIndentWhitespace);
-
-        // 2. break before arguments closing
-        $tokens->ensureWhitespaceAtIndex($endPosition + 1, 0, $this->closingBracketNewlineIndentWhitespace);
-
-        for ($i = $startPosition; $i < $endPosition; ++$i) {
-            $currentToken = $tokens[$i];
-
-            // 3. new line after each comma ",", instead of just space
-            if ($currentToken->getContent() === ',') {
-                $tokens->ensureWhitespaceAtIndex($i + 1, 0, $this->newlineIndentWhitespace);
-            }
-        }
-    }
-
-    private function getLengthFromStartEnd(int $startPosition, int $endPosition, Tokens $tokens): int
-    {
-        $lineLength = 0;
-
-        // compute from function to start of line
-        $currentPosition = $startPosition;
-        while (! Strings::startsWith($tokens[$currentPosition]->getContent(), PHP_EOL)) {
-            $lineLength += strlen($tokens[$currentPosition]->getContent());
-            --$currentPosition;
-        }
-
-        // get spaces to first line
-        $lineLength += strlen($tokens[$currentPosition]->getContent());
-
-        // get length from start of function till end of arguments - with spaces as one
-        $currentPosition = $startPosition;
-        while ($currentPosition < $endPosition) {
-            $currentToken = $tokens[$currentPosition];
-            if ($currentToken->isGivenKind(T_WHITESPACE)) {
-                ++$lineLength;
-                ++$currentPosition;
-                continue;
-            }
-
-            $lineLength += strlen($tokens[$currentPosition]->getContent());
-            ++$currentPosition;
-        }
-
-        // get length from end or arguments to first line break
-        $currentPosition = $endPosition;
-        while (! Strings::startsWith($tokens[$currentPosition]->getContent(), PHP_EOL)) {
-            $currentToken = $tokens[$currentPosition];
-
-            $lineLength += strlen($currentToken->getContent());
-            ++$currentPosition;
-        }
-
-        return $lineLength;
-    }
-
-    private function inlineItems(int $startPosition, int $endPosition, Tokens $tokens, int $currentPosition): void
-    {
-        // replace PHP_EOL with " "
-        for ($i = $currentPosition; $i < $endPosition; ++$i) {
-            $currentToken = $tokens[$i];
-
-            if (! $currentToken->isGivenKind(T_WHITESPACE)) {
-                continue;
-            }
-
-            $previousToken = $tokens[$i - 1];
-            $nextToken = $tokens[$i + 1];
-
-            // @todo make dynamic by type
-            if ($previousToken->getContent() === '(' || $nextToken->getContent() === ')') {
-                $tokens->clearAt($i);
-                continue;
-            }
-
-            $tokens[$i] = new Token([T_WHITESPACE, ' ']);
-        }
+        return Strings::startsWith($nextToken->getContent(), '<<<');
     }
 }
